data-assistant-web/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
oauth2client
db-dtypes
jinja2
google-generativeai